remove cycles
