+++ Cargo.toml
@@ -3,7 +3,7 @@ name = "bertie"
 version = "0.1.0-pre.1"
 authors = ["Karthikeyan Bhargavan <karthik@cryspen.com>"]
 license = "Apache-2.0"
-edition = "2018"
+edition = "2021"
 documentation = "https://docs.rs/bertie"
 description = "Minimal TLS 1.3 implementation"
 readme = "README.md"
